+++ src/features/profile/presentation/pages/profile-snackshop-orders.page.tsx
@@ -70,7 +70,7 @@ export function ProfileSnackshopOrders() {
     let calculatedPrice = 0;
 
     if (row.purchase_amount) {
-      calculatedPrice += parseInt(row.purchase_amount);
+      calculatedPrice += parseFloat(row.purchase_amount);
     }
 
     if (row.cod_fee) {
